Sync with git.
**Changelog:**

- Obtain gsettings transparency key value before showing the panel
 - When creating a panel, the transparency key value needs to be obtained before showing the panel. This allows transparency panel gsettings overrides.
- Resolve mutter 7 minimize animations / 4K animation
 - Mutter 7 removes the deprecated scaled-gravity animation that budgie desktop previously used.  This patch replaces this
with the recommended clutter equivalents. In addition, the window minimize/unminimize animation did not account for scaling factors so on 4K desktops the animation was broken.  This patch resolves this as well. Mutter 6 & 7 Actor performance improvements makes the existing animation effect a little 'flashy'.  This patch increases the transition speed by 30ms - so no real effect on user control, but improves the smoothness of the animation effect.
- Revert to previous filtering logic for IconTasklistApplet.